Themes: Resolved merge conflict on InstallThemeButton. See: https://github.com/Automattic/wp-calypso/pull/52776
